requirements.txt
# ============================================================
# GitLab Handbook RAG Chatbot - Dependencies
# ============================================================
# Pinned versions for Streamlit Community Cloud deployment
# ============================================================

# --- Core LLM & Embeddings ---
google-generativeai>=0.8.0
google-genai>=1.0.0

# --- Vector Database ---
chromadb>=0.5.0

# --- LangChain (RAG Orchestration) ---
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0

# --- Backend API ---
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0

# --- Frontend ---
streamlit>=1.40.0

# --- Data Collection & Processing ---
beautifulsoup4>=4.12.0
requests>=2.32.0
markdownify>=0.14.0
gitpython>=3.1.0

# --- Utilities ---
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0

# --- Testing ---
pytest>=8.0.0
httpx>=0.27.0
